requirements.txt
python-telegram-bot==20.3
pandas
openpyxl
gspread
oauth2client
python-dotenv